+++ dash/_configs.py
@@ -33,6 +33,8 @@ def load_dash_env_vars():
                 "DASH_DISABLE_VERSION_CHECK",
                 "DASH_PRUNE_ERRORS",
                 "DASH_COMPRESS",
+                "DASH_MCP_ENABLED",
+                "DASH_MCP_PATH",
                 "HOST",
                 "PORT",
             )